+++ .eslintrc.json
@@ -22,6 +22,7 @@
 	},
 	"plugins": ["react", "@typescript-eslint"],
 	"rules": {
-		"@typescript-eslint/interface-name-prefix": "off"
+		"@typescript-eslint/interface-name-prefix": "off",
+		"@typescript-eslint/explicit-function-return-type": "off"
 	}
 }
